jupyter now compatible with POST retrieve with list of ports

+++ services/dy-jupyter/handlers/input_retriever.py
@@ -6,6 +6,7 @@ import shutil
 import tempfile
 import time
 import zipfile
+from typing import List
 from pathlib import Path
 
 from notebook.base.handlers import IPythonHandler
@@ -21,7 +22,6 @@ logger = logging.getLogger(__name__)
 
 _INPUTS_FOLDER =  os.environ.get("INPUTS_FOLDER", "~/inputs")
 _OUTPUTS_FOLDER = os.environ.get("OUTPUTS_FOLDER", "~/outputs")
-
 _FILE_TYPE_PREFIX = "data:"
 _KEY_VALUE_FILE_NAME = "key_values.json"
 
@@ -62,7 +62,7 @@ async def get_time_wrapped(port):
     if isinstance(ret, Path):
         size_mb = ret.stat().st_size / 1024 / 1024
         logger.info("%s: data size: %sMB, transfer rate %sMB/s", ret.name, size_mb, size_mb / elapsed_time)    
-    return ret
+    return (port, ret)
 
 async def set_time_wrapped(port, value):
     logger.info("transfer started for %s", port.key)
@@ -75,7 +75,7 @@ async def set_time_wrapped(port, value):
         logger.info("%s: data size: %sMB, transfer rate %sMB/s", value.name, size_mb, size_mb / elapsed_time)    
     
 
-async def download_data():
+async def download_data(port_keys: List[str]) -> int:
     logger.info("retrieving data from simcore...")
     start_time = time.perf_counter()
     PORTS = node_ports.ports()
@@ -83,56 +83,69 @@ async def download_data():
     data = {}
 
     # let's gather all the data
-    download_tasks = [get_time_wrapped(port) for port in PORTS.inputs if port and port.value]
-    results = await asyncio.gather(*download_tasks)
-    logger.info("completed download %s", results)
-    # if there are files, move them to the final destination
-    # if data, copy it to a json file
-    result = iter(results)
-    for port in PORTS.inputs:
-        if not port or not port.value:
+    download_tasks = []
+    for node_input in PORTS.inputs:
+        # if port_keys contains some keys only download them
+        logger.info("Checking node %s", node_input.key)
+        if port_keys and node_input.key not in port_keys:
             continue
-        
-        value = result.__next__()
-        data[port.key] = {"key": port.key, "value": value}
+        if not node_input or node_input.value is None:            
+            continue
+        # collect coroutines
+        download_tasks.append(get_time_wrapped(node_input))
+    logger.info("retrieving %s data", len(download_tasks))    
+    
+    
+    transfer_bytes = 0
+    if download_tasks:
+        results = await asyncio.gather(*download_tasks)
+        logger.info("completed download %s", results)
+        for port, value in results:
+            data[port.key] = {"key": port.key, "value": value}
+            
+            if _FILE_TYPE_PREFIX in port.type:
+                # if there are files, move them to the final destination
+                downloaded_file = value
+                dest_path = inputs_path / port.key
+                # first cleanup
+                logger.info("removing %s", dest_path)
+                shutil.rmtree(dest_path)
+                if not downloaded_file or not downloaded_file.exists():
+                    # the link may be empty
+                    continue
+                transfer_bytes = transfer_bytes + downloaded_file.stat().st_size
+                # in case of valid file, it is either uncompressed and/or moved to the final directory
+                logger.info("creating directory %s", dest_path)
+                dest_path.mkdir(exist_ok=True, parents=True)
+                data[port.key] = {"key": port.key, "value": str(dest_path)}
 
-        if _FILE_TYPE_PREFIX in port.type:
-            downloaded_file = value
-            dest_path = inputs_path / port.key
-            # first cleanup
-            logger.info("removing %s", dest_path)
-            shutil.rmtree(dest_path)
-            if not downloaded_file:
-                # the link may be empty
-                continue
-            # in case of valid file, it is either uncompressed and/or moved to the final directory
-            logger.info("creating directory %s", dest_path)
-            dest_path.mkdir(exist_ok=True, parents=True)
-            data[port.key] = {"key": port.key, "value": str(dest_path)}
-
-            if zipfile.is_zipfile(downloaded_file):
-                logger.info("unzipping %s", downloaded_file)
-                with zipfile.ZipFile(downloaded_file) as zip_file:
-                    zip_file.extractall(dest_path, members=_no_relative_path_zip(zip_file))
-                logger.info("all unzipped in %s", dest_path)
+                if zipfile.is_zipfile(downloaded_file):
+                    logger.info("unzipping %s", downloaded_file)
+                    with zipfile.ZipFile(downloaded_file) as zip_file:
+                        zip_file.extractall(dest_path, members=_no_relative_path_zip(zip_file))
+                    logger.info("all unzipped in %s", dest_path)
+                else:
+                    logger.info("moving %s", downloaded_file)
+                    dest_path = dest_path / Path(downloaded_file).name
+                    shutil.move(downloaded_file, dest_path)
+                    logger.info("all moved to %s", dest_path)
             else:
-                logger.info("moving %s", downloaded_file)
-                dest_path = dest_path / Path(downloaded_file).name
-                shutil.move(downloaded_file, dest_path)
-                logger.info("all moved to %s", dest_path)
-
+                transfer_bytes = transfer_bytes + sys.getsizeof(value)
+    # if data other than file, copy them to a json file
     data_file = inputs_path / _KEY_VALUE_FILE_NAME
     data_file.write_text(json.dumps(data))
     stop_time = time.perf_counter()
     logger.info("all data retrieved from simcore in %sseconds: %s", stop_time - start_time, data)
+    return transfer_bytes
 
-async def upload_data():
+async def upload_data(port_keys: List[str]) -> int:
     logger.info("uploading data to simcore...")
     start_time = time.perf_counter()
     PORTS = node_ports.ports()
     outputs_path = Path(_OUTPUTS_FOLDER).expanduser()
     upload_tasks = []
     temp_files = []
+    transfer_bytes = 0
     for port in PORTS.outputs:
         logger.debug("uploading data to port '%s' with value '%s'...", port.key, port.value)
         if _FILE_TYPE_PREFIX in port.type:
@@ -167,6 +180,7 @@ async def upload_data():
                 Path(file_path).unlink()
     stop_time = time.perf_counter()
     logger.info("all data uploaded to simcore in %sseconds", stop_time-start_time)
+    return transfer_bytes
 
 
 class RetrieveHandler(IPythonHandler):
@@ -174,19 +188,49 @@ class RetrieveHandler(IPythonHandler):
         PORTS = node_ports.ports()
         _create_ports_sub_folders(PORTS.inputs, Path(_INPUTS_FOLDER).expanduser())
         _create_ports_sub_folders(PORTS.outputs, Path(_OUTPUTS_FOLDER).expanduser())
+        self.file_transfer_history = {}
 
     async def get(self):
         try:
-            await asyncio.gather(download_data(), upload_data())
+            results = await asyncio.gather(download_data(port_keys=[]), upload_data(port_keys=[]))
+            transfered_size = sum(results)
+            self.write(json.dumps({
+                "data": {
+                    "size_bytes": transfered_size
+                }
+            }))
             self.set_status(200)
         except node_ports.exceptions.NodeportsException as exc:
             logger.exception("Unexpected problem when processing retrieve call")
             self.set_status(500, reason=str(exc))
-        except Exception: #pylint: disable=broad-except
+        except Exception as exc: #pylint: disable=broad-except
             logger.exception("Unexpected problem when processing retrieve call")
-            self.set_status(500)
+            self.set_status(500, reason=str(exc))
         finally:
             self.finish('completed retrieve!')
+    
+    async def post(self):
+        request_contents = json.loads(self.request.body)
+        ports = request_contents["port_keys"]
+        logger.info("getting data of ports %s from previous node with POST request...", ports)
+        try:
+            results = await asyncio.gather(download_data(ports), upload_data(ports))
+            transfered_size = sum(results)
+            self.write(json.dumps({
+                "data": {
+                    "size_bytes": transfered_size
+                }
+            }))
+            self.set_status(200)
+        except node_ports.exceptions.NodeportsException as exc:
+            logger.exception("Unexpected problem when processing retrieve call")
+            self.set_status(500, reason=str(exc))
+        except Exception as exc: #pylint: disable=broad-except
+            logger.exception("Unexpected problem when processing retrieve call")
+            self.set_status(500, reason=str(exc))
+        finally:
+            self.finish()
+
 
 def _create_ports_sub_folders(ports: node_ports._items_list.ItemsList, parent_path: Path): # pylint: disable=protected-access
     values = {}
